WIP: move handler.go into core.go
